Final Report Upload
This is an upload of the final code and the final report including all of the modeling and data manipulation. The accompanying report is the final paper for the Rutgers MSDS program as well as the most complete report for the project. The report was cowritten with the partners (whose names appear in the code files) but as the report was for my final application only my name appears on the final paper.